requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Pillow==10.4.0
django-filter==23.5
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0